Refactor: 사용하지 않는 API 삭제
