+++ core/src/main/java/org/bookwoori/core/domain/member/controller/MemberController.java
@@ -33,13 +33,7 @@ public class MemberController {
         return ResponseEntity.ok().build();
     }
 
-    @Operation(summary = "유저 서재 조회", description = "유저의 서재를 조회합니다.")
-    @GetMapping(value = "/{memberId}/library")
-    public ResponseEntity<?> getMemberLibrary(@PathVariable("memberId") final Long memberId) {
-        return ResponseEntity.ok(memberFacade.getMemberLibrary(memberId));
-    }
-
-    @Operation(summary = "유저 프로필 조회", description = "유저의 프로필을 조회합니다.")
+    @Operation(summary = "유저 조회", description = "유저의 서재 및 프로필을 조회합니다.")
     @GetMapping(value = "/{memberId}")
     public ResponseEntity<?> getMemberProfile(@PathVariable("memberId") final Long memberId) {
         return ResponseEntity.ok(memberFacade.getMemberProfile(memberId));
