+++ core/src/main/java/org/bookwoori/core/domain/member/facade/MemberFacade.java
@@ -3,8 +3,7 @@ package org.bookwoori.core.domain.member.facade;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.bookwoori.core.domain.member.dto.request.UpdateMemberRequestDto;
-import org.bookwoori.core.domain.member.dto.response.MemberLibraryResponseDto;
-import org.bookwoori.core.domain.member.dto.response.MemberProfileResponseDto;
+import org.bookwoori.core.domain.member.dto.response.MemberResponseDto;
 import org.bookwoori.core.domain.member.entity.Member;
 import org.bookwoori.core.domain.member.service.MemberService;
 import org.bookwoori.core.global.s3.S3Util;
@@ -28,18 +27,10 @@ public class MemberFacade {
     }
 
     @Transactional(readOnly = true)
-    public MemberLibraryResponseDto getMemberLibrary(Long memberId) {
+    public MemberResponseDto getMemberProfile(Long memberId) {
         Member member = memberService.getMemberById(memberId);
         Member currentMember = memberService.getCurrentMember();
         boolean isMine = member.equals(currentMember);
-        return MemberLibraryResponseDto.from(member, isMine);
-    }
-
-    @Transactional(readOnly = true)
-    public MemberProfileResponseDto getMemberProfile(Long memberId) {
-        Member member = memberService.getMemberById(memberId);
-        Member currentMember = memberService.getCurrentMember();
-        boolean isMine = member.equals(currentMember);
-        return MemberProfileResponseDto.from(member, isMine);
+        return MemberResponseDto.from(member, isMine);
     }
 }
